requirements.txt
dbt-bigquery>=1.8.0
google-cloud-bigquery>=3.17.0
pandas>=2.0.0
